requirements.txt
# Web Framework
Flask>=2.3.0
Flask-SocketIO>=5.0.0
eventlet>=0.33.0
Flask-Limiter>=3.0.0
Flask-CORS>=4.0.0

# Computer Vision & Face Recognition
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mediapipe>=0.10.0

# NEW: InsightFace for advanced face recognition
insightface>=0.7.3
onnxruntime>=1.15.0

# NEW: Vector search
faiss-cpu>=1.7.4

# Image Processing
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Database
SQLAlchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
DBUtils>=3.0.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.0

# API Documentation
flasgger>=0.9.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
